+++ test/runtests.jl
@@ -44,6 +44,7 @@ const RUN_E2E = get(ENV, "MCP_TEST_E2E", ON_CI ? "false" : "true") == "true"
     include("protocol/parameters.jl")
     include("protocol/test_versioning.jl")
     include("protocol/test_modern.jl")
+    include("protocol/test_mrtr.jl")
     include("utils/serialization.jl")
     include("utils/logging.jl")
     include("transports/test_stdio.jl")